# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMP Stock Watcher API Documentations",
        "contact": {
            "email": "i.g.b.n.satyawibawa@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://warulaga.dapurkode.com",
            "description": "Dapurkode API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Testing Server"
        }
    ],
    "paths": {
...
        "/api/wallet/get-resource-commodity": {
            "get": {
                "tags": [
                    "Holdlist Commodity Controller",
                    "Holdlist Get Resource"
                ],
                "summary": "Commodities that avaiable to add on holdlist",
                "operationId": "holdlistCommoditiesGetResource",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Searching keyword",
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Collections of commodities",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "commodityResources": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Id of collection",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "description": "Name of stock",
                                                        "type": "string",
                                                        "example": "Gold"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                }
            }
        },
        "/api/wallet/store-commodity": {
            "post": {
                "tags": [
                    "Holdlist Commodity Controller",
                    "Holdlist Store"
                ],
                "summary": "Add commodity to holdlist",
                "operationId": "holdlistCommoditiesStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    },
                                    "amount": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        },
        "/api/wallet/remove-commodity": {
            "post": {
                "tags": [
                    "Holdlist Commodity Controller",
                    "Holdlist Remove"
                ],
                "summary": "Remove commodity from holdlist",
                "operationId": "holdlistCommoditiesRemove",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}